+++ tests/Feature/Auth/AuthenticationTest.php
@@ -48,7 +48,7 @@ class AuthenticationTest extends TestCase
         ]);
 
         $this->assertAuthenticated();
-        $response->assertRedirect(route('dashboard', absolute: false));
+        $response->assertRedirect(route('feed.index', absolute: false));
     }
 
     public function test_users_can_not_authenticate_with_invalid_password(): void
@@ -75,7 +75,7 @@ class AuthenticationTest extends TestCase
         ]);
 
         $this->assertAuthenticatedAs($user);
-        $response->assertRedirect(route('dashboard', absolute: false));
+        $response->assertRedirect(route('feed.index', absolute: false));
     }
 
     public function test_users_can_login_to_an_intended_group_page_with_existing_membership(): void